bump int sizes to int64 to support >2GB

+++ src/abf/backup_log.cpp
@@ -263,7 +263,7 @@ LogBackupFile LogBackupFile::from_value(tinyxml2::XMLElement& element, const std
     }
     XMLElement* sizeElement = element.FirstChildElement("Size");
     if (sizeElement && sizeElement->GetText()) {
-        file.Size = std::stoi(sizeElement->GetText());
+        file.Size = std::stoll(sizeElement->GetText());
     }
     XMLElement* storagePathElement = element.FirstChildElement("StoragePath");
     if (storagePathElement && storagePathElement->GetText()) {
